Add operating system check

+++ packwiz-download-latest.sh
@@ -10,7 +10,11 @@ case "$OS_TYPE" in
 		;;
 esac
 
-printf -v DOWNLOAD_LINK 'https://nightly.link/packwiz/packwiz/workflows/go/main/%s%%2064-bit%%20x86.zip' "$OS_NAME"
-wget -O "packwiz.zip" "$DOWNLOAD_LINK"
-unzip "packwiz.zip"
-rm "packwiz.zip"
+if [[ "$OS_NAME" == "" ]]; then
+	echo "Error: Cannot determine operating system."
+else
+	printf -v DOWNLOAD_LINK 'https://nightly.link/packwiz/packwiz/workflows/go/main/%s%%2064-bit%%20x86.zip' "$OS_NAME"
+	wget -O "packwiz.zip" "$DOWNLOAD_LINK"
+	unzip "packwiz.zip"
+	rm "packwiz.zip"
+fi
